qcacld-3.0: Flow Policy Manager(FPM) (nl80211 User Interface)
a. Exposes NL interface to user to configure flows.
b. Allow user to add/delete/update policies.
c. Remove cookie and use only policy_id as 64-bit
   unique identifier.

CRs-Fixed: 3692806

+++ components/dp/core/inc/wlan_fpm_table.h
@@ -32,7 +32,7 @@
  */
 struct policy_notifier_data {
 	struct flow_info flow;
-	uint32_t policy_id;
+	uint64_t policy_id;
 	uint8_t prio;
 };
 
@@ -150,7 +150,7 @@ QDF_STATUS fpm_policy_update(struct fpm_table *fpm, struct dp_policy *policy);
  *
  * Return: 0 if successful
  */
-QDF_STATUS fpm_policy_rem(struct fpm_table *fpm, uint32_t cookie);
+QDF_STATUS fpm_policy_rem(struct fpm_table *fpm, uint64_t cookie);
 
 /*
  * fpm_policy_get() - FPM policy list get
